requirements.txt
playwright>=1.44.0
playwright-stealth>=1.0.6
python-dotenv>=1.0.0
